pyproject.toml
[tool.poetry]
name = "Celline"
version = "0.1.0"
description = ""
authors = ["Yuya Sato <yu.pisces.556223@akane.waseda.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.4"
tqdm = "^4.64.1"
requests-html = "^0.10.0"
pysradb = "^1.4.2"
toml = "^0.10.2"
PyYAML = "^6.0"
argparse = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.8.0"
isort = "^5.10.1"
ipykernel = "^6.15.3"
yapf = "^0.32.0"
autopep8 = "^1.7.0"
mypy = "^0.971"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"